requirements.txt
flask==2.3.0
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-jwt-extended==4.5.2
flask-bcrypt==1.0.1
flask-mail==0.9.1
flask-cors==4.0.0
marshmallow==3.19.0
email-validator==2.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytest==7.3.1
gunicorn==20.1.0
requests==2.28.2
